requirements.txt
langchain
langchain-core
langchain-openai
langchain-ollama
langgraph
sentence-transformers
openai
python-dotenv
langchain-community
tqdm
ipywidgets
jupyter
wheel
kernel
setuptools
chardet
langchain-milvus
openpyxl
gradio


# Core frameworks
fastapi
uvicorn
gradio
python-dotenv

# LLM and Agent frameworks
groq

# LlamaIndex (for embeddings and document handling)
llama-index-core
llama-index-embeddings-gemini
llama-cloud
llama-cloud-services

# Databases and vector stores
lancedb
chromadb
duckdb
sqlalchemy

# Excel / Data handling
pandas
polars
openpyxl
fastexcel
numpy
pyarrow

# Google Gemini + API clients
google-generativeai



requests
tqdm
tenacity
coloredlogs
pydantic
typing-extensions
protobuf
orjson
aiohttp
anyio
httpx
aiofiles

# Optional visualization & misc
matplotlib-inline
rich


tabulate
